fix(manager): fix some bug)
