+++ manager/param/add_source.go
@@ -1,0 +1,22 @@
+package param
+
+import "time"
+
+type AddSourceRequest struct {
+	Name        string
+	Description string
+	ProjectID   string
+	OwnerID     string
+}
+
+type AddSourceResponse struct {
+	WriteKey    string
+	Name        string
+	Description string
+	ProjectID   string
+	OwnerID     string
+	Status      bool
+	CreateAt    time.Time
+	UpdateAt    time.Time
+	DeleteAt    *time.Time
+}
