+++ manager/param/update_source.go
@@ -4,25 +4,6 @@ import (
 	"time"
 )
 
-type AddSourceRequest struct {
-	Name        string
-	Description string
-	ProjectID   string
-	OwnerID     string
-}
-
-type AddSourceResponse struct {
-	WriteKey    string
-	Name        string
-	Description string
-	ProjectID   string
-	OwnerID     string
-	Status      bool
-	CreateAt    time.Time
-	UpdateAt    time.Time
-	DeleteAt    *time.Time
-}
-
 type UpdateSourceRequest struct {
 	Name        string
 	Description string
