+++ manager/service/sourceservice/delete.go
@@ -1,0 +1,13 @@
+package sourceservice
+
+import (
+	"github.com/ormushq/ormus/pkg/richerror"
+)
+
+func (s *Service) DeleteSource(id string) error {
+	if err := s.repo.DeleteSource(id); err != nil {
+		return richerror.New("DeleteSource").WhitWarpError(err)
+	}
+
+	return nil
+}
