+++ manager/service/sourceservice/service.go
@@ -1,10 +1,13 @@
 package sourceservice
 
-import "github.com/ormushq/ormus/manager/entity"
+import (
+	"github.com/ormushq/ormus/manager/entity"
+	"github.com/ormushq/ormus/manager/param"
+)
 
 type SourceRepo interface {
-	InsertSource(source *entity.Source) error
-	UpdateSource(id string, source *entity.Source) error
+	InsertSource(source *entity.Source) (*param.AddSourceResponse, error)
+	UpdateSource(id string, source *entity.Source) (*param.UpdateSourceResponse, error)
 	DeleteSource(id string) error
 	GetUserSourceById(ownerID, id string) (*entity.Source, error)
 	IsSourceAlreadyCreatedByName(name string) (bool, error)
